change indices' types from float to int

+++ models/est.py
@@ -124,11 +124,11 @@ class QuantizationLayer(nn.Module):
 
         p = (p+1)/2  # maps polarity to 0, 1
 
-        idx_before_bins = x \
-                          + W * y \
+        idx_before_bins = x.int() \
+                          + W * y.int()\
                           + 0 \
-                          + W * H * C * p \
-                          + W * H * C * 2 * b
+                          + W * H * C * p.int() \
+                          + W * H * C * 2 * b.int()
 
         for i_bin in range(C):
             values = t * self.value_layer.forward(t-i_bin/(C-1))
